Fix tests after merge
