+++ src/Chart.ts
@@ -1,7 +1,19 @@
 import {Chart, ChartConfiguration, ChartOptions} from "chart.js";
 import {DataFilterer, DataFiltererOptions, FilteredData} from "./DataFilterer";
-import {touchstoneYears} from "./Dictionaries";
+//import {touchstoneYears} from "./Dictionaries";
 import {plotColours} from "./PlotColours";
+
+export const touchstoneYears: { [code: string]: number} = {
+    "201210gavi-201303gavi" : 2011,
+    "201210gavi-201810gavi" : 2017,
+    "201310gavi"            : 2012,
+    "201310gavi-201403gavi" : 2012,
+    "201310gavi-201810gavi" : 2017,
+    "201510gavi"            : 2014,
+    "201510gavi-201810gavi" : 2017,
+    "201710gavi"            : 2016,
+    "201710gavi-201810gavi" : 2017,
+};
 
 export interface CustomChartOptions extends DataFiltererOptions {
     plotTitle: string;
