+++ src/Dictionaries.ts
@@ -47,15 +47,3 @@ export function diseaseCodeToName(disCode: string) {
 
 export const diseaseVaccineLookup: {[code: string]: string[]} =
                         loadObjectFromJSONFile("./diseaseVaccines.json");
-
-export const touchstoneYears: { [code: string]: number} = {
-    "201210gavi-201303gavi" : 2011,
-    "201210gavi-201810gavi" : 2017,
-    "201310gavi"            : 2012,
-    "201310gavi-201403gavi" : 2012,
-    "201310gavi-201810gavi" : 2017,
-    "201510gavi"            : 2014,
-    "201510gavi-201810gavi" : 2017,
-    "201710gavi"            : 2016,
-    "201710gavi-201810gavi" : 2017,
-};
